requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.51.0
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-taggit==3.0.0
email-reply-parser==0.5.12
filelock==3.7.0
gunicorn==20.1.0
idna==3.3
Markdown==3.4.1
mysql-connector-python==8.0.29
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.2.0
